demo-mcp/requirements.txt
fastmcp
mcp
langchain-core
langchain-openai
langgraph
python-dotenv
psycopg2-binary
